add server simple test
