+++ web/package.json
@@ -7,6 +7,7 @@
     "build": "astro build",
     "preview": "astro preview",
     "start": "node server",
+    "test": "node --test",
     "astro": "astro",
     "lint": "biome lint"
   },
